Check if active order present against product before delete

+++ controllers/product.js
@@ -23,8 +23,8 @@ exports.create = (req, res) => {
 // support update on 1 product
 exports.update = (req, res) => {
     Product.updateOne({ name: req.body.name }, {
-            $set: { price: req.body.price, stock: req.body.stock }
-        })
+        $set: { price: req.body.price, stock: req.body.stock }
+    })
         .then((result) => {
             res.status(STATUS.CREATED)
                 .send({
@@ -42,15 +42,43 @@ exports.update = (req, res) => {
 // support delete of multiple products by passing in array
 exports.deleteProducts = (req, res) => {
 
+    // First check if there are active orders for the product to be deleted
+    const queryExp = [];
+    const productsToDelete = req.body;
+    productsToDelete.forEach(product => {
+        queryExp.push({
+            name: product,
+            order: 0
+        })
+    })
 
-
-    Product.deleteMany({ name: { "$in": [...req.body] } })
+    Product.find({ $or: queryExp })
         .then((result) => {
-            res.status(STATUS.OK)
-                .send({
-                    message: "Products deleted successfully",
-                    details: result
+            if (!result) {
+                res.status(STATUS.INTERNAL_SERVER_ERROR).send({
+                    message: "Internal Server Error",
+                    details: "Some unknown error has occured"
                 });
+            } else if (result && result.length !== req.body.length) {
+                res.status(STATUS.INTERNAL_SERVER_ERROR).send({
+                    message: "Product deletion error",
+                    details: "Either product to be deleted not found, or there is an active order that prevents product from deletion"
+                });
+            } else {
+                Product.deleteMany({ name: { "$in": [...req.body] } })
+                    .then((result) => {
+                        res.status(STATUS.OK)
+                            .send({
+                                message: "Products deleted successfully",
+                                details: result
+                            });
+                    }).catch((err) => {
+                        res.status(STATUS.INTERNAL_SERVER_ERROR)
+                            .send({
+                                message: err
+                            });
+                    });
+            }
         }).catch((err) => {
             res.status(STATUS.INTERNAL_SERVER_ERROR)
                 .send({
